separate the tooltip tokenid component and adjust row spaces
